VC - View 분리
