+++ CommerceApp/Feature/Purchase/PurchaseViewController.swift
@@ -10,134 +10,26 @@ import Combine
 
 final class PurchaseViewController: UIViewController {
     private var viewModel: PurchaseViewModel = PurchaseViewModel()
-    private var scrollViewConstraints: [NSLayoutConstraint]?
-    private var titleLabelConstraints: [NSLayoutConstraint]?
-    private var purchaseItemStackViewConstraints: [NSLayoutConstraint]?
-    private var purchaseButtonConstraints: [NSLayoutConstraint]?
     private var subscriptions = Set<AnyCancellable>()
+    private var rootView: PurchaseRootView = PurchaseRootView()
     
-    private var scrollView: UIScrollView = {
-        let view: UIScrollView = UIScrollView()
-        view.alwaysBounceVertical = true
-        view.translatesAutoresizingMaskIntoConstraints = false
-        return view
-    }()
-    
-    private var containerView: UIView = {
-        let view: UIView = UIView()
-        view.translatesAutoresizingMaskIntoConstraints = false
-        return view
-    }()
-    
-    private var titleLabel: UILabel = {
-        let label: UILabel = UILabel()
-        label.text = "주문 상품 목록"
-        label.font = CPFont.m17
-        label.textColor = CPColor.UIKit.bk
-        label.translatesAutoresizingMaskIntoConstraints = false
-        return label
-    }()
-    
-    private var purchaseItemStackView: UIStackView = {
-        let view: UIStackView = UIStackView()
-        view.axis = .vertical
-        view.alignment = .fill
-        view.distribution = .fill
-        view.spacing = 7
-        view.translatesAutoresizingMaskIntoConstraints = false
-        return view
-    }()
-    
-    private var purchaseButton: UIButton = {
-        let btn: UIButton = UIButton()
-        btn.setTitle("결제하기", for: .normal)
-        btn.titleLabel?.font = CPFont.m16
-        btn.layer.backgroundColor = CPColor.UIKit.keyColorBlue.cgColor
-        btn.layer.cornerRadius = 5
-        btn.translatesAutoresizingMaskIntoConstraints = false
-        return btn
-    }()
+    override func loadView() {
+        view = rootView
+    }
     
     override func viewDidLoad() {
         super.viewDidLoad()
         view.backgroundColor = .white
-        addSubView()
         bindViewModel()
         viewModel.process(.loadData)
     }
-    
-    override func updateViewConstraints() {
-        if scrollViewConstraints == nil {
-            let constraints = [
-                scrollView.topAnchor.constraint(equalTo: view.safeAreaLayoutGuide.topAnchor),
-                scrollView.leadingAnchor.constraint(equalTo: view.safeAreaLayoutGuide.leadingAnchor),
-                scrollView.trailingAnchor.constraint(equalTo: view.safeAreaLayoutGuide.trailingAnchor),
-                scrollView.bottomAnchor.constraint(equalTo: purchaseButton.topAnchor),
-                
-                containerView.topAnchor.constraint(equalTo: scrollView.topAnchor),
-                containerView.leadingAnchor.constraint(equalTo: scrollView.leadingAnchor),
-                containerView.trailingAnchor.constraint(equalTo: scrollView.trailingAnchor),
-                containerView.bottomAnchor.constraint(equalTo: scrollView.bottomAnchor),
-                containerView.widthAnchor.constraint(equalTo: scrollView.widthAnchor)
-            ]
-            
-            NSLayoutConstraint.activate(constraints)
-            scrollViewConstraints = constraints
-        }
-        
-        if titleLabelConstraints == nil,
-           let superView = titleLabel.superview {
-            let constraints = [
-                titleLabel.topAnchor.constraint(equalTo: superView.topAnchor, constant: 33),
-                titleLabel.leadingAnchor.constraint(equalTo: superView.leadingAnchor, constant: 33),
-                titleLabel.trailingAnchor.constraint(equalTo: superView.trailingAnchor, constant: -33)
-            ]
-            NSLayoutConstraint.activate(constraints)
-            titleLabelConstraints = constraints
-        }
-        
-        if purchaseItemStackViewConstraints == nil {
-            let constraints = [
-                purchaseItemStackView.topAnchor.constraint(equalTo: titleLabel.bottomAnchor, constant: 19),
-                purchaseItemStackView.leadingAnchor.constraint(equalTo: titleLabel.leadingAnchor, constant: 20),
-                purchaseItemStackView.trailingAnchor.constraint(equalTo: titleLabel.trailingAnchor, constant: -20),
-                purchaseItemStackView.bottomAnchor.constraint(equalTo: titleLabel.bottomAnchor, constant: 33)
-            ]
-            NSLayoutConstraint.activate(constraints)
-            purchaseItemStackViewConstraints = constraints
-        }
-        
-        if purchaseButtonConstraints == nil {
-            let constraints = [
-                purchaseButton.leadingAnchor.constraint(equalTo: view.safeAreaLayoutGuide.leadingAnchor, constant: 20),
-                purchaseButton.trailingAnchor.constraint(equalTo: view.safeAreaLayoutGuide.trailingAnchor, constant: -20),
-                purchaseButton.bottomAnchor.constraint(equalTo: view.safeAreaLayoutGuide.bottomAnchor, constant: -32),
-                purchaseButton.heightAnchor.constraint(equalToConstant: 50)
-            ]
-            NSLayoutConstraint.activate(constraints)
-            titleLabelConstraints = constraints
-        }
-        super.updateViewConstraints()
-    }
-    
-    private func addSubView() {
-        view.addSubview(scrollView)
-        scrollView.addSubview(containerView)
-        containerView.addSubview(titleLabel)
-        containerView.addSubview(purchaseItemStackView)
-        view.addSubview(purchaseButton)
-    }
-    
+
     private func bindViewModel() {
         viewModel.$state
             .receive(on: DispatchQueue.main)
             .sink { [weak self] _ in
-                self?.purchaseItemStackView.subviews.forEach {
-                    $0.removeFromSuperview()
-                }
-                self?.viewModel.state.purchaseItems?.forEach {
-                    self?.purchaseItemStackView.addArrangedSubview(PurchaseSelectedItemView(viewModel: .init(title: $0.title, description: $0.description)))
-                }
+                guard let viewModels = self?.viewModel.state.purchaseItems else { return }
+                self?.rootView.setPurchaseItemStack(viewModels)
             }
             .store(in: &subscriptions)
     }
